requirements.txt
# --- Web Framework ---
fastapi==0.115.2
uvicorn[standard]==0.30.6

# --- HTTP Client ---
httpx==0.27.2

# --- File Upload & Processing ---
python-multipart==0.0.9
python-magic==0.4.27
Pillow==10.4.0
reportlab==4.2.2
img2pdf==0.5.1
pypdf==3.8.0

# --- Testing ---
pytest
pytest-asyncio
